Update example.html
This is a simple reference example for the group project, that incorporates a lot of the learnt material in the presentation. In their groups they can create an article each and place it in a section for a simple webpage. I was planning on doing multiple pages but I'll add it as a further reading reference as it would confuse many people new to HTML when running it locally.